Add tests for DummyTransformer
